Adding html elements and tags
