+++ flaskr/static/css/dochine.css
@@ -1,0 +1,3 @@
+.contrib {
+    color: blueviolet;
+}
